backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.2.0
gunicorn==21.2.0
